feat: migrate non-map pages to TypeScript with typed hooks and state management
