+++ src/pages/Home.tsx
@@ -1,11 +1,11 @@
 const Home = () => {
   return (
     <>
-      <div className='pagebox'>
+      <div className="pagebox">
         <main>
           <h1>Welcome to the homepage!</h1>
           <p>You can do this, I believe in you.</p>
-        </main>       
+        </main>
       </div>
     </>
   );
