--- src/pages/LoginRegister.jsx
@@ -1,73 +1,0 @@
-import { useEffect, useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { useDispatch, useSelector } from 'react-redux';
-import { loginUser, registerUser } from '../features/authentication/authSlice';
-import './LoginRegister.css';
-
-const LoginRegister = () => {
-    const dispatch = useDispatch();
-    const { loading, error } = useSelector((state) => state.auth);
-
-    const [formData, setFormData] = useState({ username: '', password: '' });
-    const [isRegister, setIsRegister] = useState(false);
-
-    const handleChange = (e) => {
-        setFormData((prev) => ({ ...prev, [e.target.name]: e.target.value }));
-    };
-
-    const handleSubmit = (e) => {
-        e.preventDefault();
-        if (isRegister) {
-            dispatch(registerUser(formData));
-        } else {
-            dispatch(loginUser(formData));
-        }
-    };
-
-    const navigate = useNavigate();
-    const { user } = useSelector((state) => state.auth);
-
-    useEffect(() => {
-        if (user) {
-            navigate('/');
-        }
-    }, [user, navigate]);
-
-    return (
-        <div className="login-container">
-            <h2>{isRegister ? 'Register' : 'Login'}</h2>
-            <form onSubmit={handleSubmit} className="login-form">
-                <label htmlFor="username">Username</label>
-                <input
-                    type="text"
-                    name="username"
-                    placeholder="Username"
-                    value={formData.username}
-                    onChange={handleChange}
-                    required
-                />
-                <label htmlFor="password">Password</label>
-                <input
-                    type="password"
-                    name="password"
-                    placeholder="Password"
-                    value={formData.password}
-                    onChange={handleChange}
-                    required
-                />
-                {error && <p className="error">{error}</p>}
-                <button type="submit" disabled={loading}>
-                    {loading ? 'Please wait...' : isRegister ? 'Register' : 'Login'}
-                </button>
-            </form>
-            <div className="toggle-mode">
-                {isRegister ? 'Already have an account? ' : "Don't have an account? "}
-                <span onClick={() => setIsRegister(!isRegister)}>
-                    {isRegister ? 'Login' : 'Register'}
-                </span>
-            </div>
-        </div>
-    );
-};
-
-export default LoginRegister;
