+++ src/pages/LoginRegister.tsx
@@ -1,0 +1,80 @@
+import { useEffect, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { useAppDispatch, useAppSelector } from '../app/hooks';
+import type { RootState } from '../app/store';
+import { loginUser, registerUser } from '../features/authentication/authSlice';
+import './LoginRegister.css';
+
+interface FormData {
+  username: string;
+  password: string;
+}
+
+const LoginRegister = () => {
+  const dispatch = useAppDispatch();
+  const { loading, error } = useAppSelector((state: RootState) => state.auth);
+
+  const [formData, setFormData] = useState<FormData>({ username: '', password: '' });
+  const [isRegister, setIsRegister] = useState(false);
+
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setFormData((prev) => ({ ...prev, [e.target.name]: e.target.value }));
+  };
+
+  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    if (isRegister) {
+      dispatch(registerUser(formData));
+    } else {
+      dispatch(loginUser(formData));
+    }
+  };
+
+  const navigate = useNavigate();
+  const { user } = useAppSelector((state: RootState) => state.auth);
+
+  useEffect(() => {
+    if (user) {
+      navigate('/');
+    }
+  }, [user, navigate]);
+
+  return (
+    <div className="login-container">
+      <h2>{isRegister ? 'Register' : 'Login'}</h2>
+      <form onSubmit={handleSubmit} className="login-form">
+        <label htmlFor="username">Username</label>
+        <input
+          type="text"
+          name="username"
+          placeholder="Username"
+          value={formData.username}
+          onChange={handleChange}
+          required
+        />
+        <label htmlFor="password">Password</label>
+        <input
+          type="password"
+          name="password"
+          placeholder="Password"
+          value={formData.password}
+          onChange={handleChange}
+          required
+        />
+        {error && <p className="error">{error}</p>}
+        <button type="submit" disabled={loading}>
+          {loading ? 'Please wait...' : isRegister ? 'Register' : 'Login'}
+        </button>
+      </form>
+      <div className="toggle-mode">
+        {isRegister ? 'Already have an account? ' : "Don't have an account? "}
+        <span onClick={() => setIsRegister(!isRegister)}>
+          {isRegister ? 'Login' : 'Register'}
+        </span>
+      </div>
+    </div>
+  );
+};
+
+export default LoginRegister;
+
